+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/search/CertifiedProductSearchDAOImpl.java
@@ -46,8 +46,12 @@ public class CertifiedProductSearchDAOImpl extends BaseDAOImpl implements Certif
 			result.setDeveloper(dbResult.getDeveloper());
 			result.setProduct(dbResult.getProduct());
 			result.setVersion(dbResult.getVersion());
+			result.setNumMeaningfulUse(dbResult.getMeaningfulUseUserCount());
+			result.setDecertificationDate(dbResult.getDecertificationDate() == null ? null : dbResult.getDecertificationDate().getTime());
 			result.setCertificationDate(dbResult.getCertificationDate().getTime());
 			result.setCertificationStatus(dbResult.getCertificationStatus());
+			result.setTransparencyAttestationUrl(dbResult.getTransparencyAttestationUrl());
+			result.setApiDocumentation(dbResult.getApiDocumentation());
 			result.setSurveillanceCount(dbResult.getSurveillanceCount());
 			result.setOpenNonconformityCount(dbResult.getOpenNonconformityCount());
 			result.setClosedNonconformityCount(dbResult.getClosedNonconformityCount());