+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/CertificationStatusEvent.java
@@ -13,15 +13,27 @@ import javax.xml.bind.annotation.XmlType;
 public class CertificationStatusEvent implements Serializable {
 	private static final long serialVersionUID = -2498656549844148886L;
 	
+	/**
+	 * Internal ID
+	 */
 	@XmlElement(required = true)
 	private Long id;
 	
+	/**
+	 * The date on which a change of certification status occurred.
+	 */
 	@XmlElement(required = true)
 	private Long eventDate;
 	
+	/**
+	 * Internal certification status ID.
+	 */
 	@XmlElement(required = true)
 	private Long certificationStatusId;
 	
+	/**
+	 * Certification status name.
+	 */
 	@XmlElement(required = true)
 	private String certificationStatusName;