+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/web/controller/CertifiedProductControllerTest.java
@@ -17,6 +17,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import org.json.JSONException;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;