+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/web/controller/SearchViewControllerTest.java
@@ -1,5 +1,6 @@
 package gov.healthit.chpl.web.controller;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -264,6 +265,29 @@ public class SearchViewControllerTest extends TestCase {
 		SearchResponse searchResponse = new SearchResponse();
 		searchResponse = searchViewController.advancedSearch(searchFilters);
 		assertTrue("searchViewController.simpleSearch() should return a SearchResponse with records", searchResponse.getRecordCount() > 0);
+	}
+	
+	/** Description: Tests that the advancedSearch returns valid SearchResponse records when refined by certification start and end date
+	 * 
+	 * Expected Result: Completes without error and returns some SearchResponse records
+	 */
+	@Transactional
+	@Test
+	public void test_advancedSearch_refineByCertificationStartAndEndDate_CompletesWithoutError() 
+			throws EntityRetrievalException, JsonProcessingException, EntityCreationException,
+			InvalidArgumentsException {
+		SecurityContextHolder.getContext().setAuthentication(adminUser);
+		SearchRequest searchFilters = new SearchRequest();
+		searchFilters.setPageNumber(0);
+		searchFilters.setPageSize(50);
+		searchFilters.setOrderBy("developer");
+		searchFilters.setSortDescending(true);
+		searchFilters.setCertificationDateStart("2015-08-01");
+		searchFilters.setCertificationDateEnd("2015-08-31");
+		
+		SearchResponse searchResponse = new SearchResponse();
+		searchResponse = searchViewController.advancedSearch(searchFilters);
+		assertEquals(Integer.valueOf(12), Integer.valueOf(searchResponse.getRecordCount()));
 	}
 	
 	@Transactional 
@@ -594,46 +618,6 @@ public class SearchViewControllerTest extends TestCase {
 				resp.getNumMeaningfulUse() == 12);
 	}
 	
-	/** 
-	 * Given that a user with no security calls the API
-	 * When the API is called at /decertifications/certified_products
-	 * Then the API returns a SearchResponse object with only decertified CPs
-	 * Then the pageSize is equivalent to the sum of all the decertified CPs
-	 * @throws EntityRetrievalException 
-	 */
-	@Transactional
-	@Test
-	public void test_searchDecertifiedCPs() throws EntityRetrievalException {	
-		SearchResponse resp = searchViewController.getDecertifiedCertifiedProducts(null, null, null, null);		
-		
-		assertTrue(resp.getResults().size() == 1);
-		assertEquals((Integer) resp.getResults().size(), resp.getPageSize());
-		for(CertifiedProductSearchResult cp : resp.getResults()){
-			assertTrue(cp.getCertificationStatus().containsValue(String.valueOf(CertificationStatusType.WithdrawnByAcb)) || 
-					cp.getCertificationStatus().containsValue(String.valueOf(CertificationStatusType.WithdrawnByDeveloperUnderReview)) ||
-					cp.getCertificationStatus().containsValue(String.valueOf(CertificationStatusType.TerminatedByOnc)));
-		}
-	}
-	
-	/** 
-	 * Given that a user with no security calls the API
-	 * When the API is called at /decertifications/certified_products
-	 * Then the API returns a SearchResponse object with only decertified inactive certificate CPs
-	 * Then the pageSize is equivalent to the sum of all the decertified CPs
-	 * @throws EntityRetrievalException 
-	 */
-	@Transactional
-	@Test
-	public void test_searchDecertifiedInactiveCertCPs() throws EntityRetrievalException {	
-		SearchResponse resp = searchViewController.getDecertifiedInactiveCertificateCertifiedProducts(null, null, null, null);		
-		
-		assertTrue(resp.getResults().size() == 6);
-		assertEquals((Integer) resp.getResults().size(), resp.getPageSize());
-		for(CertifiedProductSearchResult cp : resp.getResults()){
-			assertTrue(cp.getCertificationStatus().containsValue(String.valueOf(CertificationStatusType.WithdrawnByDeveloper)));
-		}
-	}
-	
 	@Transactional
 	@Test
 	public void testBasicSearchDefaultViewHasRequiredFields() throws JsonProcessingException, EntityRetrievalException {
@@ -666,6 +650,8 @@ public class SearchViewControllerTest extends TestCase {
 			assertNotNull(result.getVersion());
 			assertNotNull(result.getCertificationDate());
 			assertNotNull(result.getCertificationStatus());
+			assertNull(result.getDecertificationDate());
+			assertNull(result.getNumMeaningfulUse());
 			assertNotNull(result.getSurveillanceCount());
 			assertNotNull(result.getOpenNonconformityCount());
 			assertNotNull(result.getClosedNonconformityCount());
@@ -714,6 +700,8 @@ public class SearchViewControllerTest extends TestCase {
 			assertNull(result.getVersion());
 			assertNull(result.getCertificationDate());
 			assertNull(result.getCertificationStatus());
+			assertNull(result.getDecertificationDate());
+			assertNull(result.getNumMeaningfulUse());
 			assertNull(result.getSurveillanceCount());
 			assertNull(result.getOpenNonconformityCount());
 			assertNull(result.getClosedNonconformityCount());
@@ -754,6 +742,8 @@ public class SearchViewControllerTest extends TestCase {
 			assertNull(result.getVersion());
 			assertNull(result.getCertificationDate());
 			assertNull(result.getCertificationStatus());
+			assertNull(result.getDecertificationDate());
+			assertNull(result.getNumMeaningfulUse());
 			assertNull(result.getSurveillanceCount());
 			assertNull(result.getOpenNonconformityCount());
 			assertNull(result.getClosedNonconformityCount());
@@ -793,6 +783,8 @@ public class SearchViewControllerTest extends TestCase {
 			assertNull(result.getVersion());
 			assertNull(result.getCertificationDate());
 			assertNull(result.getCertificationStatus());
+			assertNull(result.getDecertificationDate());
+			assertNull(result.getNumMeaningfulUse());
 			assertNull(result.getSurveillanceCount());
 			assertNull(result.getOpenNonconformityCount());
 			assertNull(result.getClosedNonconformityCount());