+++ RELEASE_NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+## Version 11.0.0
+_19 June 2017_
+
+### Major Features
+_Backwards compatibility breaking changes_
+* Removed /decertifications/certified_products and /decertifications/inactive_certificates API calls. This same data can now be accessed through the /certified_products API method.
+
+### Minor Features
+* Add decertification date, numMeaningfulUse, transparency attestation url, and api documentation url fields to the flat search results
+* Fill in blank cells for subelements in nonconformtiy and surveillance downloads.
+
+---
+
 ## Version 10.2.0
 _5 June 2017_